hospital-ops-backend/inventory_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-alloydb-connector[pg8000]==1.12.1
sqlalchemy==2.0.23
pydantic==2.5.2